# quads/swagger.yaml
openapi: 3.0.0
info:
  version: '3.0.0'
  title: 'QUADS'
  description: 'QUADS automates the future scheduling, end-to-end provisioning and delivery of bare-metal servers and networks.'
  license:
    name: GPLv3
servers:
  - url: 'https://quads.example.com/api/v3/'
paths:
...
  /hosts/:
    get:
      summary: 'Returns a list with all the defined hosts'
      tags:
        - hosts
      responses:
        '200':
          description: 'Host name'
          headers:
            x-next:
              description: A link to the next page of responses
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Host'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Add a new host
      tags:
        - hosts
      requestBody:
        description: Body request with *host* details
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Host'
          application/xml:
            schema:
              $ref: '#/components/schemas/Host'
          text/plain:
            schema:
              type: string
      responses:
        '201':
          description: Created
  /hosts/{hostName}:
    get:
      summary: 'Returns a host by name'
      tags:
        - hosts
      parameters:
        - name: hostName
          in: path
          description: Host name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Host name'
          headers:
            x-next:
              description: A link to the next page of responses
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Host'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      summary: 'Update an existing host'
      tags:
        - hosts
      parameters:
        - name: hostName
          in: path
          description: Host name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Host name'
          headers:
            x-next:
              description: A link to the next page of responses
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Host'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      description: Delete host by host name
      tags:
        - hosts
      parameters:
        - name: hostName
          in: path
          description: Hostname
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Deleted Host'
...
  /assignments/:
    get:
      summary: 'Returns a list with all the defined assignments'
      tags:
        - assignments
      responses:
        '200':
          description: 'Assignment'
          headers:
            x-next:
              description: A link to the next page of responses
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assignment'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Add a new assignment
      tags:
        - assignments
      requestBody:
        description: Body request with *assignment* details
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Assignment'
          application/xml:
            schema:
              $ref: '#/components/schemas/Assignment'
          text/plain:
            schema:
              type: string
      security:
        - BearerAuth: [ ]
      responses:
        '201':
          description: Created
  /assignments/{assignment_id}:
    get:
      summary: 'Returns an assignment by id'
      tags:
        - assignments
      parameters:
        - name: assignment_id
          in: path
          description: Assignment ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Assignment'
          headers:
            x-next:
              description: A link to the next page of responses
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assignment'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      description: Delete assignment by id
      tags:
        - assignments
      parameters:
        - name: assignment_id
          in: path
          description: Assignment ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Deleted assignment'
...
components:
  schemas:
...
    Disk:
      type: object
      properties:
        disk_type:
          type: string
          example: nvme
        size_gb:
          type: integer
          example: 2000
        count:
          type: integer
          example: 10
        host_id:
          type: integer
          example: 10
    Memory:
      type: object
      required:
        - size_gb
        - handle
        - host_id
      properties:
        id:
          type: integer
          example: 12
        size_gb:
          type: integer
          example: 2000
        handle:
          type: string
          example: MMD
        host_id:
          type: integer
          example: 10
    Processor:
      type: object
      properties:
        id:
          type: integer
          example: 12
        handle:
          type: string
          example: CoreI7
        vendor:
          type: string
          example: Intel
        product:
          type: string
          example: i7
        cores:
          type: integer
          example: 20
        threads:
          type: integer
          example: 20
        host_id:
          type: integer
          example: 10
    Interface:
      type: object
      properties:
        id:
          type: integer
          example: 12
        name:
          type: string
          example: em1
        bios_id:
          type: string
          example: nic1
        mac_address:
          type: string
          example: aa:00:bb:11:cc:22
        switch_ip:
          type: string
          example: 10.1.1.18
        switch_port:
          type: string
          example: xt-0-0/1
        speed:
          type: integer
          example: 1000
        vendor:
          type: string
          example: Intel
        pxe_boot:
          type: boolean
          example: true
        maintenance:
          type: boolean
          example: false
        host_id:
          type: integer
          example: 10
    Host:
      type: object
      properties:
        id:
          type: integer
          example: 12
        name:
          type: string
          example: host.example.com
        model:
          type: string
          example: r640
        host_type:
          type: string
          example: vendor
        build:
          type: boolean
          example: true
        validated:
          type: boolean
          example: true
        switch_config_applied:
          type: boolean
          example: true
        broken:
          type: boolean
          example: true
        retired:
          type: boolean
          example: true
        last_build:
          type: string
          example: 2022-02-02
        cloud_id:
          type: integer
          example: 12
        default_cloud_id:
          type: integer
          example: 1
        interfaces:
          type: array
          items:
            $ref: '#/components/schemas/Interface'
        disks:
          type: array
          items:
            $ref: '#/components/schemas/Disk'
        memory:
          type: array
          items:
            $ref: '#/components/schemas/Memory'
        processors:
          type: array
          items:
            $ref: '#/components/schemas/Processor'
    Assignment:
      type: object
      properties:
        active:
          type: boolean
          example: true
        provisioned:
          type: boolean
          example: true
        validated:
          type: boolean
          example: true
        description:
          type: string
          example: Short description here
        owner:
          type: string
          example: jsbach
        ticket:
          type: string
          example: 2122
        qinq:
          type: integer
          example: 1
        wipe:
          type: boolean
          example: true
        ccuser:
          type: array
          items:
            type: string
          example: [edroste, kgodel]
        cloud_id:
          type: integer
          example: 12
        vlan_id:
          type: integer
          example: 12
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer